Update Application.php
This change removes the old captcha and adds the newer style no captcha recaptch method. I have also changed the file extensions to the included pages to make use of the newer .php file extension system of phpVMS. 

One issue is the email it sends. it is in very bad shape. The bonus is, it is sending. The mail function needs work. that is my next project.